requirements.txt
asgiref==3.2.7
astroid
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap-modal-forms==1.5.0
django-bootstrap4==1.1.1
django-braces==1.13.0
django-debug-toolbar==2.1
django-widget-tweaks==1.4.5
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
misaka==2.1.1
psycopg2-binary==2.8.4
pycparser==2.20
pygeoif==0.7
pylint==2.4.4
python-dateutil==2.6.1
python-decouple==3.3
pytz==2018.3
requests==2.22.0
sgp4==1.4
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
tle2czml==0.2
urllib3==1.25.7
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.11.2
